refactoring step_4 to smaller elements
